reorganize shared fixtures into conftest
